# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  AdminPageData-internal_transport_http_admin_AuditLogResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/AuditLogResponse'
        type: array
      total:
        type: integer
    required:
    - results
    - total
    type: object
  AdminPageData-internal_transport_http_admin_AuthEventResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/AuthEventResponse'
        type: array
      total:
        type: integer
    required:
    - results
    - total
    type: object
  AdminPageData-internal_transport_http_admin_ConversationEventResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/ConversationEventResponse'
        type: array
      total:
        type: integer
    required:
    - results
    - total
    type: object
  AdminPageData-internal_transport_http_admin_SystemEventResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/SystemEventResponse'
        type: array
      total:
        type: integer
    required:
    - results
    - total
    type: object
  AdminPageData-internal_transport_http_admin_UserResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/AdminUserResponse'
        type: array
      total:
        type: integer
    required:
    - results
    - total
    type: object
  AdminUserResponse:
    properties:
      appearancePreferences:
        type: string
      avatarURL:
        type: string
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        type: integer
      lastActiveAt:
        type: string
      lastLoginAt:
        type: string
      locale:
        type: string
      profilePreferences:
        type: string
      publicID:
        type: string
      role:
        type: string
      status:
        type: string
      timezone:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    required:
    - appearancePreferences
    - avatarURL
    - createdAt
    - displayName
    - email
    - id
    - lastActiveAt
    - lastLoginAt
    - locale
    - profilePreferences
    - publicID
    - role
    - status
    - timezone
    - updatedAt
    - username
    type: object
...
  AuditLogListResponseDoc:
    properties:
      data:
        $ref: '#/definitions/AdminPageData-internal_transport_http_admin_AuditLogResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  AuditLogResponse:
    properties:
      action:
        type: string
      actorDisplayName:
        type: string
      actorLabel:
        type: string
      actorUserID:
        type: integer
      actorUsername:
        type: string
      createdAt:
        type: string
      detailJSON:
        type: string
      id:
        type: integer
      ip:
        type: string
      requestID:
        type: string
      resource:
        type: string
      resourceID:
        type: string
      updatedAt:
        type: string
      userAgent:
        type: string
    required:
    - action
    - actorDisplayName
    - actorLabel
    - actorUserID
    - actorUsername
    - createdAt
    - detailJSON
    - id
    - ip
    - requestID
    - resource
    - resourceID
    - updatedAt
    - userAgent
    type: object
...
  AuthEventListResponseDoc:
    properties:
      data:
        $ref: '#/definitions/AdminPageData-internal_transport_http_admin_AuthEventResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  AuthEventResponse:
    properties:
      clientIP:
        type: string
      createdAt:
        type: string
      detailJSON:
        type: string
      eventType:
        type: string
      id:
        type: integer
      occurredAt:
        type: string
      reason:
        type: string
      requestID:
        type: string
      result:
        type: string
      updatedAt:
        type: string
      userAgent:
        type: string
      userDisplayName:
        type: string
      userID:
        type: integer
      userLabel:
        type: string
      username:
        type: string
    required:
    - clientIP
    - createdAt
    - detailJSON
    - eventType
    - id
    - occurredAt
    - reason
    - requestID
    - result
    - updatedAt
    - userAgent
    - userDisplayName
    - userID
    - userLabel
    - username
    type: object
...
  CleanupConversationRunsRequest:
    properties:
      runIDs:
        items:
          type: string
        maxItems: 100
        minItems: 1
        type: array
    required:
    - runIDs
    type: object
  CleanupConversationRunsResponse:
    properties:
      deletedCount:
        type: integer
      runCount:
        type: integer
    required:
    - deletedCount
    - runCount
    type: object
  CleanupConversationRunsResponseDoc:
    properties:
      data:
        $ref: '#/definitions/CleanupConversationRunsResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  CleanupLogsRequest:
    properties:
      before:
        type: string
      type:
        type: string
    required:
    - before
    - type
    type: object
  CleanupLogsResponse:
    properties:
      before:
        type: string
      deletedCount:
        type: integer
      type:
        type: string
    required:
    - before
    - deletedCount
    - type
    type: object
  CleanupLogsResponseDoc:
    properties:
      data:
        $ref: '#/definitions/CleanupLogsResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  ConversationEventDetailResponseDoc:
    properties:
      data:
        $ref: '#/definitions/ConversationEventResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  ConversationEventListResponseDoc:
    properties:
      data:
        $ref: '#/definitions/AdminPageData-internal_transport_http_admin_ConversationEventResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  ConversationEventResponse:
    properties:
      contentMarkdown:
        type: string
      conversationID:
        type: integer
      createdAt:
        type: string
      endedAt:
        type: string
      eventID:
        type: string
      eventScope:
        type: string
      eventType:
        type: string
      id:
        type: integer
      latencyMS:
        type: integer
      messageID:
        type: integer
      payloadJSON:
        type: string
      payloadOmitted:
        type: boolean
      phase:
        type: string
      runID:
        type: string
      seq:
        type: integer
      stage:
        type: string
      startedAt:
        type: string
      status:
        type: string
      summary:
        type: string
      title:
        type: string
      toolName:
        type: string
      updatedAt:
        type: string
      userDisplayName:
        type: string
      userID:
        type: integer
      userLabel:
        type: string
      username:
        type: string
    required:
    - contentMarkdown
    - conversationID
    - createdAt
    - endedAt
    - eventID
    - eventScope
    - eventType
    - id
    - latencyMS
    - messageID
    - payloadJSON
    - payloadOmitted
    - phase
    - runID
    - seq
    - stage
    - startedAt
    - status
    - summary
    - title
    - toolName
    - updatedAt
    - userDisplayName
    - userID
    - userLabel
    - username
    type: object
...
  PatchUserRequest:
    properties:
      avatarURL:
        type: string
      displayName:
        type: string
      locale:
        type: string
      profilePreferences:
        type: string
      reason:
        type: string
      timezone:
        type: string
    type: object
...
  RevokeUserSessionsResponse:
    properties:
      revoked:
        type: boolean
    required:
    - revoked
    type: object
  RevokeUserSessionsResponseDoc:
    properties:
      data:
        $ref: '#/definitions/RevokeUserSessionsResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
  SystemEventListResponseDoc:
    properties:
      data:
        $ref: '#/definitions/AdminPageData-internal_transport_http_admin_SystemEventResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  SystemEventResponse:
    properties:
      createdAt:
        type: string
      detailJSON:
        type: string
      event:
        type: string
      id:
        type: integer
      level:
        type: string
      message:
        type: string
      requestID:
        type: string
      resource:
        type: string
      resourceID:
        type: string
      source:
        type: string
      traceID:
        type: string
      updatedAt:
        type: string
    required:
    - createdAt
    - detailJSON
    - event
    - id
    - level
    - message
    - requestID
    - resource
    - resourceID
    - source
    - traceID
    - updatedAt
    type: object
...
  UserDataResponse:
    properties:
      user:
        $ref: '#/definitions/AdminUserResponse'
    required:
    - user
    type: object
  UserDataResponseDoc:
    properties:
      data:
        $ref: '#/definitions/UserDataResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
  UserListResponseDoc:
    properties:
      data:
        $ref: '#/definitions/AdminPageData-internal_transport_http_admin_UserResponse'
      errorMsg:
        type: string
    required:
    - data
    - errorMsg
    type: object
...
info:
  contact: {}
  description: DEEIX Chat 后端 API 文档
  title: DEEIX Chat API
  version: "0.4.0"
paths:
...
  /admin/audit-logs:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AuditLogListResponseDoc'
      security:
      - BearerAuth: []
      summary: List audit logs
      tags:
      - admin
  /admin/conversation-events:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ConversationEventListResponseDoc'
      security:
      - BearerAuth: []
      summary: List conversation events
      tags:
      - admin
  /admin/conversation-events/{id}:
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ConversationEventDetailResponseDoc'
      security:
      - BearerAuth: []
      summary: Get a conversation event
      tags:
      - admin
  /admin/conversation-events/cleanup:
    post:
      parameters:
      - description: Cleanup request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CleanupConversationRunsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CleanupConversationRunsResponseDoc'
      security:
      - BearerAuth: []
      summary: Remove conversation event runs
      tags:
      - admin
...
  /admin/logs/cleanup:
    post:
      parameters:
      - description: Cleanup request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/CleanupLogsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CleanupLogsResponseDoc'
      security:
      - BearerAuth: []
      summary: Remove nonfinancial logs before a timestamp
      tags:
      - admin
...
  /admin/system-events:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SystemEventListResponseDoc'
      security:
      - BearerAuth: []
      summary: List system events
      tags:
      - admin
  /admin/user-auth-events:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AuthEventListResponseDoc'
      security:
      - BearerAuth: []
      summary: List authentication events
      tags:
      - admin
  /admin/users:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserListResponseDoc'
      security:
      - BearerAuth: []
      summary: List principals
      tags:
      - admin
  /admin/users/{id}:
    patch:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Profile fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/PatchUserRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserDataResponseDoc'
      security:
      - BearerAuth: []
      summary: Patch a principal profile
      tags:
      - admin
  /admin/users/{id}/revoke-sessions:
    post:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RevokeUserSessionsResponseDoc'
      security:
      - BearerAuth: []
      summary: Revoke local sessions
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"